Renames display to terminal
